Move the renamed classes to more suitable places.
BFD-652
